Fixes a possible memory leak in the useMatchMedia custom hook
